Revert "Allow concurrent-safe open() and fopen() in frontend code for Windows"
This reverts commit f02259fe93e75d5443a2fabe2f2f38b81924ab36, in the
v11 branch only.

The hack this required in initdb.c should probably have clued us that it
wasn't really ready, but we didn't get the hint.  Subsequent developments
have made clear that it affected text-vs-binary behavior in a lot of
places, and there's no reason to think that any of those behavioral changes
are desirable.  There's no time to fix this before 11beta4, so just revert
for the moment.  We can keep working on this in HEAD, and maybe reconsider
a back-patch once we're satisfied things are stable.

(I take the blame for this fiasco, having encouraged Michael to back-patch
a change at the last possible moment before beta wrap.)